+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Allow overriding a person's bio on the person plugin
+- Add new section on the course detail page to display related programs
 - Display "code" on the course detail page and allow frontend editing
 
 ### Changed